Implement expire-after-write functionality properly, add abstract builder, start adding non-reactive proxy manager.
